requirements.txt
argon2-cffi==18.3.0
bcrypt==3.1.4
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.1.3
django-allauth==0.38.0
django-heroku==0.3.1
django-tagging==0.4.6
Faker==1.0.0
gunicorn==19.9.0
idna==2.7
oauthlib==2.1.0
Pillow==5.3.0
psycopg2==2.7.6.1
pycparser==2.19
python-dateutil==2.7.5
python3-openid==3.1.0
pytz==2018.7
requests==2.20.1
requests-oauthlib==1.0.0
six==1.11.0
text-unidecode==1.2
urllib3==1.24.1
whitenoise==4.1.1
